[add] separando as categorias e os movimentos por usuario

+++ EasyBudget/EasyBudget/Repositories/Implementations/CategoryRepository.cs
@@ -15,64 +15,48 @@ namespace EasyBudget.Repositories.Implementations
             _conn = conn;
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(int id, int userId)
         {
             string command = @"
 DELETE FROM Category
 WHERE id = @Id
+AND user_id = @UserId
 ";
 
-            await _conn.ExecuteAsync(command, new { Id = id });
+            await _conn.ExecuteAsync(command, new { Id = id, UserId = userId });
         }
 
-        public async Task<bool> ExistsByIdAsync(int id)
+        public async Task<bool> ExistsByIdAsync(int id, int userId)
         {
             string command = @"
 SELECT
     CASE
         WHEN EXISTS(
-            SELECT 1 FROM Category WHERE id = @Id)
+            SELECT 1 FROM Category WHERE id = @Id AND user_id = @UserId)
         THEN 1
         ELSE 0
     END;
 ";
 
-            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(command, new { Id = id }));
+            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(command, new { Id = id, UserId = userId }));
         }
 
-        public async Task<bool> ExistsByNameAndTypeAsync(string name, FinancialType type)
+        public async Task<bool> ExistsByNameAndTypeAsync(string name, FinancialType type, int userId)
         {
             string command = @"
 SELECT
     CASE
         WHEN EXISTS(
-            SELECT 1 FROM Category WHERE name = @Name and type = @Type)
+            SELECT 1 FROM Category WHERE name = @Name AND type = @Type AND user_id = @UserId)
         THEN 1
         ELSE 0
     END;
 ";
 
-            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(command, new { Name = name, Type = type }));
+            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(command, new { Name = name, Type = type, UserId = userId }));
         }
 
-        public async Task<IEnumerable<ReadCategoryDto>> FindAllAsync()
-        {
-            string query = @$"
-SELECT
-    id as Id,
-    name as Name,
-    type as Type,
-    CASE
-        WHEN type = {(int)FinancialType.Income}
-            THEN '{FinancialType.Income}'
-        ELSE '{FinancialType.Expense}'
-    END as TypeName
-FROM Category";
-
-            return await _conn.QueryAsync<ReadCategoryDto>(query);
-        }
-
-        public async Task<ReadCategoryDto?> FindByIdAsync(int id)
+        public async Task<IEnumerable<ReadCategoryDto>> FindAllAsync(int userId)
         {
             string query = @$"
 SELECT
@@ -85,16 +69,35 @@ SELECT
         ELSE '{FinancialType.Expense}'
     END as TypeName
 FROM Category
-WHERE id = @Id;";
+WHERE user_id = @UserId";
 
-            return await _conn.QueryFirstOrDefaultAsync<ReadCategoryDto?>(query, new { Id = id });
+            return await _conn.QueryAsync<ReadCategoryDto>(query, new { UserId = userId });
+        }
+
+        public async Task<ReadCategoryDto?> FindByIdAsync(int id, int userId)
+        {
+            string query = @$"
+SELECT
+    id as Id,
+    name as Name,
+    type as Type,
+    CASE
+        WHEN type = {(int)FinancialType.Income}
+            THEN '{FinancialType.Income}'
+        ELSE '{FinancialType.Expense}'
+    END as TypeName
+FROM Category
+WHERE id = @Id
+AND user_id = @UserId;";
+
+            return await _conn.QueryFirstOrDefaultAsync<ReadCategoryDto?>(query, new { Id = id, UserId = userId });
         }
 
         public async Task<ReadCategoryDto> InsertAsync(CreateCategoryDto category)
         {
             string insert = @$"
 INSERT INTO Category
-(name, type)
+(name, type, user_id)
 OUTPUT
     INSERTED.id as Id,
     INSERTED.name as Name,
@@ -105,7 +108,7 @@ OUTPUT
         ELSE '{FinancialType.Expense}'
     END as TypeName
 VALUES
-(@Name, @Type) ;
+(@Name, @Type, @UserId) ;
 ";
 
             return await _conn.QuerySingleAsync<ReadCategoryDto>(insert, category);
@@ -119,6 +122,7 @@ UPDATE Category
         name = @Name,
         type = @Type
 WHERE id = @Id
+AND user_id = @UserId
 ";
             await _conn.ExecuteAsync(update, category);
         }
