+++ EasyBudget/EasyBudget/Repositories/Implementations/MovementRepository.cs
@@ -16,51 +16,18 @@ namespace EasyBudget.Repositories.Implementations
             _conn = conn;
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(int id, int userId)
         {
             string command = @"
 DELETE FROM Movement
 WHERE id = @Id
+AND user_id = @UserId
 ";
 
-            await _conn.ExecuteAsync(command, new { Id = id });
+            await _conn.ExecuteAsync(command, new { Id = id, UserId = userId });
         }
 
-        public async Task<IEnumerable<ReadMovementDto>> FindAllAsync()
-        {
-            string query = @$"
-SELECT
-    [m].id as Id,
-    [m].amount as Amount,
-    [m].title as Title,
-    [m].date as Date,
-    [m].type as Type,
-    CASE
-        WHEN [m].type = {(int)FinancialType.Income}
-            THEN '{FinancialType.Income}'
-        ELSE '{FinancialType.Expense}'
-    END as TypeName,
-    [m].description as Description,
-    [c].id as Id,
-    [c].name as Name,
-    [c].type as Type,
-    CASE
-        WHEN [c].type = {(int)FinancialType.Income}
-            THEN '{FinancialType.Income}'
-        ELSE '{FinancialType.Expense}'
-    END as TypeName
-FROM Movement as [m]
-    JOIN Category as [c]
-        on [m].category_id = [c].id;";
-
-            return await _conn.QueryAsync<ReadMovementDto, ReadCategoryDto, ReadMovementDto>(query, (movement, category) =>
-            {
-                movement.Category = category;
-                return movement;
-            });
-        }
-
-        public async Task<IEnumerable<ReadMovementDto>> FindAllByCategoryAsync(int categoryId)
+        public async Task<IEnumerable<ReadMovementDto>> FindAllAsync(int userId)
         {
             string query = @$"
 SELECT
@@ -86,16 +53,16 @@ SELECT
 FROM Movement as [m]
     JOIN Category as [c]
         on [m].category_id = [c].id
-WHERE [m].category_id = @CategoryId;";
+WHERE [m].user_id = @UserId;";
 
             return await _conn.QueryAsync<ReadMovementDto, ReadCategoryDto, ReadMovementDto>(query, (movement, category) =>
             {
                 movement.Category = category;
                 return movement;
-            }, new { CategoryId = categoryId });
+            }, new { UserId = userId });
         }
 
-        public async Task<ReadMovementDto?> FindByIdAsync(int id)
+        public async Task<IEnumerable<ReadMovementDto>> FindAllByCategoryAsync(int categoryId, int userId)
         {
             string query = @$"
 SELECT
@@ -121,21 +88,58 @@ SELECT
 FROM Movement as [m]
     JOIN Category as [c]
         on [m].category_id = [c].id
-WHERE [m].id = @Id;";
+WHERE [m].category_id = @CategoryId
+AND [m].user_id = @UserId;";
+
+            return await _conn.QueryAsync<ReadMovementDto, ReadCategoryDto, ReadMovementDto>(query, (movement, category) =>
+            {
+                movement.Category = category;
+                return movement;
+            }, new { CategoryId = categoryId, UserId =  userId});
+        }
+
+        public async Task<ReadMovementDto?> FindByIdAsync(int id, int userId)
+        {
+            string query = @$"
+SELECT
+    [m].id as Id,
+    [m].amount as Amount,
+    [m].title as Title,
+    [m].date as Date,
+    [m].type as Type,
+    CASE
+        WHEN [m].type = {(int)FinancialType.Income}
+            THEN '{FinancialType.Income}'
+        ELSE '{FinancialType.Expense}'
+    END as TypeName,
+    [m].description as Description,
+    [c].id as Id,
+    [c].name as Name,
+    [c].type as Type,
+    CASE
+        WHEN [c].type = {(int)FinancialType.Income}
+            THEN '{FinancialType.Income}'
+        ELSE '{FinancialType.Expense}'
+    END as TypeName
+FROM Movement as [m]
+    JOIN Category as [c]
+        on [m].category_id = [c].id
+WHERE [m].id = @Id
+AND [m].user_id = @UserId;";
 
             return (await _conn.QueryAsync<ReadMovementDto, ReadCategoryDto, ReadMovementDto?>(query, (movement, category) =>
             {
                 movement.Category = category;
                 return movement;
-            }, new { Id = id })).FirstOrDefault();
+            }, new { Id = id, UserId = userId })).FirstOrDefault();
         }
 
         public async Task<ReadMovementDto> InsertAsync(CreateMovementDto movement)
         {
             string insert = @$"
 INSERT INTO Movement
-(amount, title, date, type, category_id, description) VALUES
-(@Amount, @Title, @Date, @Type, @CategoryId, @Description);
+(amount, title, date, type, category_id, description, user_id) VALUES
+(@Amount, @Title, @Date, @Type, @CategoryId, @Description, @UserId)
 
 SELECT
     [m].id as Id,
@@ -170,7 +174,7 @@ WHERE [m].id = SCOPE_IDENTITY();
             }, param: movement)).First();
         }
 
-        public async Task<decimal> SumAllMovementsAmount()
+        public async Task<decimal> SumAllMovementsAmount(int userId)
         {
             string query = $@"
 SELECT
@@ -181,9 +185,10 @@ SELECT
             ELSE amount
         END
     )
-FROM Movement;
+FROM Movement
+WHERE user_id = @UserId;
 ";
-            return decimal.Round(await _conn.QueryFirstAsync<decimal>(query), 2);
+            return decimal.Round(await _conn.QueryFirstAsync<decimal>(query, new { UserId = userId }), 2);
         }
 
         public async Task UpdateAsync(UpdateMovementDto movement)
@@ -198,6 +203,7 @@ UPDATE Movement
         category_id = @CategoryId,
         description = @Description
 WHERE id = @Id
+AND user_id = @UserId
 ";
 
             await _conn.ExecuteAsync(update, movement);
@@ -215,35 +221,37 @@ UPDATE Movement
         category_id = @CategoryId,
         description = @Description
 WHERE id = @Id
+AND user_id = @UserId
 ";
 
             await _conn.ExecuteAsync(update, movements);
         }
 
-        public async Task ReplaceCategory(int oldCategoryId, int newCategoryId)
+        public async Task ReplaceCategory(int oldCategoryId, int newCategoryId, int userId)
         {
             string update = @"
 UPDATE Movement
     SET category_id = @NewCategoryId
-WHERE category_id = @OldCategoryId;
+WHERE category_id = @OldCategoryId
+AND user_id = @UserId;
 ";
 
-            await _conn.ExecuteAsync(update, new { NewCategoryId = newCategoryId, OldCategoryId = oldCategoryId });
+            await _conn.ExecuteAsync(update, new { NewCategoryId = newCategoryId, OldCategoryId = oldCategoryId, UserId = userId });
         }
 
-        public async Task<bool> ExistsByIdAsync(int id)
+        public async Task<bool> ExistsByIdAsync(int id, int userId)
         {
             string query = @"
 SELECT
     CASE
         WHEN EXISTS(
-            SELECT 1 FROM Movement WHERE id = @Id)
+            SELECT 1 FROM Movement WHERE id = @Id AND user_id = @UserId)
         THEN 1
         ELSE 0
     END;
 ";
 
-            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(query, new { Id = id }));
+            return Convert.ToBoolean(await _conn.ExecuteScalarAsync<int>(query, new { Id = id, UserId = userId }));
         }
     }
 }
